feat: 좋아요 상태 업데이트 함수 분리
