+++ src/features/feed-main-like/consts/index.ts
@@ -1,0 +1,6 @@
+import { FetchFeeds } from '@/shared/consts';
+
+export interface FeedsQueryData {
+  queryParams: number[];
+  pages: FetchFeeds[];
+}
